fix: move function to helpers
